Fixed compile warning.

+++ src/Helpers.cpp
@@ -419,7 +419,7 @@ bool ofxImGui::AddRadio(ofParameter<int>& parameter, vector<string> labels, int 
 	ImGui::PushID(parameter.getName().c_str());
 	{
 		ImGui::Columns(columns);
-		for (int i = 0; i < labels.size(); ++i)
+		for (size_t i = 0; i < labels.size(); ++i)
 		{
 			result |= ImGui::RadioButton(GetUniqueName(labels[i]), &tmpRef, i);
 			ImGui::NextColumn();
@@ -565,7 +565,7 @@ bool ofxImGui::AddRange(const string& name, ofParameter<glm::vec4>& parameterMin
 bool ofxImGui::AddValues(const string& name, vector<glm::tvec2<int>>& values, int minValue, int maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderInt2(GetUniqueName(iname), glm::value_ptr(values[i]), minValue, maxValue);
@@ -577,7 +577,7 @@ bool ofxImGui::AddValues(const string& name, vector<glm::tvec2<int>>& values, in
 bool ofxImGui::AddValues(const string& name, vector<glm::tvec3<int>>& values, int minValue, int maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderInt3(GetUniqueName(iname), glm::value_ptr(values[i]), minValue, maxValue);
@@ -589,7 +589,7 @@ bool ofxImGui::AddValues(const string& name, vector<glm::tvec3<int>>& values, in
 bool ofxImGui::AddValues(const string& name, vector<glm::tvec4<int>>& values, int minValue, int maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderInt4(GetUniqueName(iname), glm::value_ptr(values[i]), minValue, maxValue);
@@ -601,7 +601,7 @@ bool ofxImGui::AddValues(const string& name, vector<glm::tvec4<int>>& values, in
 bool ofxImGui::AddValues(const string& name, vector<glm::vec2>& values, float minValue, float maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderFloat2(GetUniqueName(iname), glm::value_ptr(values[i]), minValue, maxValue);
@@ -613,7 +613,7 @@ bool ofxImGui::AddValues(const string& name, vector<glm::vec2>& values, float mi
 bool ofxImGui::AddValues(const string& name, vector<glm::vec3>& values, float minValue, float maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderFloat3(GetUniqueName(iname), glm::value_ptr(values[i]), minValue, maxValue);
@@ -625,7 +625,7 @@ bool ofxImGui::AddValues(const string& name, vector<glm::vec3>& values, float mi
 bool ofxImGui::AddValues(const string& name, vector<glm::vec4>& values, float minValue, float maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderFloat4(GetUniqueName(iname), glm::value_ptr(values[i]), minValue, maxValue);
@@ -639,7 +639,7 @@ bool ofxImGui::AddValues(const string& name, vector<glm::vec4>& values, float mi
 bool ofxImGui::AddValues(const string& name, vector<ofVec2f>& values, float minValue, float maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderFloat2(GetUniqueName(iname), values[i].getPtr(), minValue, maxValue);
@@ -651,7 +651,7 @@ bool ofxImGui::AddValues(const string& name, vector<ofVec2f>& values, float minV
 bool ofxImGui::AddValues(const string& name, vector<ofVec3f>& values, float minValue, float maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderFloat3(GetUniqueName(iname), values[i].getPtr(), minValue, maxValue);
@@ -663,7 +663,7 @@ bool ofxImGui::AddValues(const string& name, vector<ofVec3f>& values, float minV
 bool ofxImGui::AddValues(const string& name, vector<ofVec4f>& values, float minValue, float maxValue)
 {
 	auto result = false;
-	for (int i = 0; i < values.size(); ++i)
+	for (size_t i = 0; i < values.size(); ++i)
 	{
 		const auto iname = name + " " + ofToString(i);
 		result |= ImGui::SliderFloat4(GetUniqueName(iname), values[i].getPtr(), minValue, maxValue);
